+++ FlsTaleQuiz/Business/Models/UserAnswer.cs
@@ -2,8 +2,8 @@
 {
     public class UserAnswer
     {
-        public long QuestionId { get; set; }
+        public int QuestionId { get; set; }
 
-        public long AnswerId { get; set; }
+        public int AnswerId { get; set; }
     }
 }
